feat: whiteboard, macos
