+++ src/whiteboard/windows.rs
@@ -1,72 +1,19 @@
-use crate::ipc::{self, new_listener, Connection, Data};
-use hbb_common::{
-    allow_err,
-    anyhow::anyhow,
-    bail, log, sleep,
-    tokio::{
-        self,
-        sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender},
-        time::interval_at,
-    },
-    ResultType,
-};
-use lazy_static::lazy_static;
-use serde_derive::{Deserialize, Serialize};
+use super::{server::EVENT_PROXY, Cursor, CustomEvent};
+use hbb_common::{anyhow::anyhow, bail, log, ResultType};
 use softbuffer::{Context, Surface};
-use std::{
-    collections::HashMap,
-    num::NonZeroU32,
-    sync::{Arc, RwLock},
-    time::Instant,
-};
+use std::{collections::HashMap, num::NonZeroU32, sync::Arc, time::Instant};
 #[cfg(target_os = "linux")]
 use tao::platform::unix::WindowBuilderExtUnix;
 #[cfg(target_os = "windows")]
 use tao::platform::windows::WindowBuilderExtWindows;
 use tao::{
+    dpi::{PhysicalPosition, PhysicalSize},
     event::{Event, WindowEvent},
-    event_loop::{ControlFlow, EventLoopBuilder, EventLoopProxy},
+    event_loop::{ControlFlow, EventLoopBuilder},
     window::WindowBuilder,
 };
 use tiny_skia::{Color, FillRule, Paint, PathBuilder, PixmapMut, Point, Stroke, Transform};
 use ttf_parser::Face;
-
-lazy_static! {
-    static ref EVENT_PROXY: RwLock<Option<EventLoopProxy<(String, CustomEvent)>>> =
-        RwLock::new(None);
-    static ref TX_WHITEBOARD: RwLock<Option<UnboundedSender<(String, CustomEvent)>>> =
-        RwLock::new(None);
-    static ref CONNS: RwLock<HashMap<String, Conn>> = Default::default();
-}
-
-struct Conn {
-    last_cursor_pos: (f32, f32), // For click ripple
-    last_cursor_evt: LastCursorEvent,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "t", content = "c")]
-pub enum CustomEvent {
-    Cursor(Cursor),
-    Clear,
-    Exit,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "t")]
-pub struct Cursor {
-    pub x: f32,
-    pub y: f32,
-    pub argb: u32,
-    pub btns: i32,
-    pub text: String,
-}
-
-struct LastCursorEvent {
-    evt: Option<CustomEvent>,
-    tm: Instant,
-    c: usize,
-}
 
 // A helper struct to bridge `ttf-parser` and `tiny-skia`.
 struct PathBuilderWrapper<'a> {
@@ -148,314 +95,6 @@ fn draw_text(
     }
 }
 
-#[inline]
-pub fn get_key_cursor(conn_id: i32) -> String {
-    format!("{}-cursor", conn_id)
-}
-
-pub fn register_whiteboard(k: String) {
-    std::thread::spawn(|| {
-        allow_err!(start_whiteboard_());
-    });
-    let mut conns = CONNS.write().unwrap();
-    if !conns.contains_key(&k) {
-        conns.insert(
-            k,
-            Conn {
-                last_cursor_pos: (0.0, 0.0),
-                last_cursor_evt: LastCursorEvent {
-                    evt: None,
-                    tm: Instant::now(),
-                    c: 0,
-                },
-            },
-        );
-    }
-}
-
-pub fn unregister_whiteboard(k: String) {
-    let mut conns = CONNS.write().unwrap();
-    conns.remove(&k);
-    let is_conns_empty = conns.is_empty();
-    drop(conns);
-
-    TX_WHITEBOARD.read().unwrap().as_ref().map(|tx| {
-        allow_err!(tx.send((k, CustomEvent::Clear)));
-    });
-    if is_conns_empty {
-        std::thread::spawn(|| {
-            let mut whiteboard = TX_WHITEBOARD.write().unwrap();
-            whiteboard.as_ref().map(|tx| {
-                allow_err!(tx.send(("".to_string(), CustomEvent::Exit)));
-                // Simple sleep to wait the whiteboard process exiting.
-                std::thread::sleep(std::time::Duration::from_millis(3_00));
-            });
-            whiteboard.take();
-        });
-    }
-}
-
-pub fn update_whiteboard(k: String, e: CustomEvent) {
-    let mut conns = CONNS.write().unwrap();
-    let Some(conn) = conns.get_mut(&k) else {
-        return;
-    };
-    match &e {
-        CustomEvent::Cursor(cursor) => {
-            conn.last_cursor_evt.c += 1;
-            conn.last_cursor_evt.tm = Instant::now();
-            if cursor.btns == 0 {
-                // Send one movement event every 4.
-                if conn.last_cursor_evt.c > 3 {
-                    conn.last_cursor_evt.c = 0;
-                    conn.last_cursor_evt.evt = None;
-                    tx_send_event(conn, k, e);
-                } else {
-                    conn.last_cursor_evt.evt = Some(e);
-                }
-            } else {
-                if let Some(evt) = conn.last_cursor_evt.evt.take() {
-                    tx_send_event(conn, k.clone(), evt);
-                    conn.last_cursor_evt.c = 0;
-                }
-                let click_evt = CustomEvent::Cursor(Cursor {
-                    x: conn.last_cursor_pos.0,
-                    y: conn.last_cursor_pos.1,
-                    argb: cursor.argb,
-                    btns: cursor.btns,
-                    text: cursor.text.clone(),
-                });
-                tx_send_event(conn, k, click_evt);
-            }
-        }
-        _ => {
-            tx_send_event(conn, k, e);
-        }
-    }
-}
-
-#[inline]
-fn tx_send_event(conn: &mut Conn, k: String, event: CustomEvent) {
-    if let CustomEvent::Cursor(cursor) = &event {
-        if cursor.btns == 0 {
-            conn.last_cursor_pos = (cursor.x, cursor.y);
-        }
-    }
-
-    TX_WHITEBOARD.read().unwrap().as_ref().map(|tx| {
-        allow_err!(tx.send((k, event)));
-    });
-}
-
-#[tokio::main(flavor = "current_thread")]
-async fn start_whiteboard_() -> ResultType<()> {
-    let mut tx_whiteboard = TX_WHITEBOARD.write().unwrap();
-    if tx_whiteboard.is_some() {
-        log::warn!("Whiteboard already started");
-        return Ok(());
-    }
-
-    loop {
-        if !crate::platform::is_prelogin() {
-            break;
-        }
-        sleep(1.).await;
-    }
-    let mut stream = None;
-    if let Ok(s) = ipc::connect(1000, "_whiteboard").await {
-        stream = Some(s);
-    } else {
-        #[allow(unused_mut)]
-        #[allow(unused_assignments)]
-        let mut args = vec!["--whiteboard"];
-        #[allow(unused_mut)]
-        #[cfg(target_os = "linux")]
-        let mut user = None;
-
-        let run_done;
-        if crate::platform::is_root() {
-            let mut res = Ok(None);
-            for _ in 0..10 {
-                #[cfg(not(any(target_os = "linux")))]
-                {
-                    log::debug!("Start whiteboard");
-                    res = crate::platform::run_as_user(args.clone());
-                }
-                #[cfg(target_os = "linux")]
-                {
-                    log::debug!("Start whiteboard");
-                    res = crate::platform::run_as_user(
-                        args.clone(),
-                        user.clone(),
-                        None::<(&str, &str)>,
-                    );
-                }
-                if res.is_ok() {
-                    break;
-                }
-                log::error!("Failed to run whiteboard: {res:?}");
-                sleep(1.).await;
-            }
-            if let Some(task) = res? {
-                super::CHILD_PROCESS.lock().unwrap().push(task);
-            }
-            run_done = true;
-        } else {
-            run_done = false;
-        }
-        if !run_done {
-            log::debug!("Start whiteboard");
-            super::CHILD_PROCESS
-                .lock()
-                .unwrap()
-                .push(crate::run_me(args)?);
-        }
-        for _ in 0..20 {
-            sleep(0.3).await;
-            if let Ok(s) = ipc::connect(1000, "_whiteboard").await {
-                stream = Some(s);
-                break;
-            }
-        }
-        if stream.is_none() {
-            bail!("Failed to connect to connection manager");
-        }
-    }
-
-    let mut stream = stream.ok_or(anyhow!("none stream"))?;
-    let (tx, mut rx) = unbounded_channel();
-    tx_whiteboard.replace(tx);
-    drop(tx_whiteboard);
-    let _call_on_ret = crate::common::SimpleCallOnReturn {
-        b: true,
-        f: Box::new(move || {
-            let _ = TX_WHITEBOARD.write().unwrap().take();
-        }),
-    };
-
-    let dur = tokio::time::Duration::from_millis(300);
-    let mut timer = interval_at(tokio::time::Instant::now() + dur, dur);
-    timer.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-    loop {
-        tokio::select! {
-            res = rx.recv() => {
-                match res {
-                    Some(data) => {
-                        if matches!(data.1, CustomEvent::Exit) {
-                            break;
-                        } else {
-                            allow_err!(stream.send(&Data::Whiteboard(data)).await);
-                            timer.reset();
-                        }
-                    }
-                    None => {
-                        bail!("expected");
-                    }
-                }
-            },
-            _ = timer.tick() => {
-                let mut conns = CONNS.write().unwrap();
-                for (k, conn) in conns.iter_mut() {
-                    if conn.last_cursor_evt.tm.elapsed().as_millis() > 300 {
-                        if let Some(evt) = conn.last_cursor_evt.evt.take() {
-                            allow_err!(stream.send(&Data::Whiteboard((k.clone(), evt))).await);
-                            conn.last_cursor_evt.c = 0;
-                        }
-                    }
-                }
-            }
-        }
-    }
-    allow_err!(
-        stream
-            .send(&Data::Whiteboard(("".to_string(), CustomEvent::Exit)))
-            .await
-    );
-    Ok(())
-}
-
-pub fn run() {
-    let (tx_exit, rx_exit) = unbounded_channel();
-    std::thread::spawn(move || {
-        start_ipc(rx_exit);
-    });
-    if let Err(e) = create_event_loop() {
-        log::error!("Failed to create event loop: {}", e);
-        tx_exit.send(()).ok();
-        return;
-    }
-}
-
-#[tokio::main(flavor = "current_thread")]
-async fn start_ipc(mut rx_exit: UnboundedReceiver<()>) {
-    match new_listener("_whiteboard").await {
-        Ok(mut incoming) => loop {
-            tokio::select! {
-                _ = rx_exit.recv() => {
-                    log::info!("Exiting IPC");
-                    break;
-                }
-                res = incoming.next() => match res {
-                    Some(result) => match result {
-                        Ok(stream) => {
-                            log::debug!("Got new connection");
-                            tokio::spawn(handle_new_stream(Connection::new(stream)));
-                        }
-                        Err(err) => {
-                            log::error!("Couldn't get whiteboard client: {:?}", err);
-                        }
-                    },
-                    None => {
-                        log::error!("Failed to get whiteboard client");
-                    }
-                }
-            }
-        },
-        Err(err) => {
-            log::error!("Failed to start whiteboard ipc server: {}", err);
-        }
-    }
-}
-
-async fn handle_new_stream(mut conn: Connection) {
-    loop {
-        tokio::select! {
-            res = conn.next() => {
-                match res {
-                    Err(err) => {
-                        log::info!("whiteboard ipc connection closed: {}", err);
-                        break;
-                    }
-                    Ok(Some(data)) => {
-                        match data {
-                            Data::Whiteboard((k, evt)) => {
-                                if matches!(evt, CustomEvent::Exit) {
-                                    log::info!("whiteboard ipc connection closed");
-                                    break;
-                                } else {
-                                    EVENT_PROXY.read().unwrap().as_ref().map(|ep| {
-                                        allow_err!(ep.send_event((k, evt)));
-                                    });
-                                }
-                            }
-                            _ => {
-
-                            }
-                        }
-                    }
-                    Ok(None) => {
-                        log::info!("whiteboard ipc connection closed");
-                        break;
-                    }
-                }
-            }
-        }
-    }
-    EVENT_PROXY.read().unwrap().as_ref().map(|ep| {
-        allow_err!(ep.send_event(("".to_string(), CustomEvent::Exit)));
-    });
-}
-
 fn create_font_face() -> ResultType<Face<'static>> {
     let mut font_db = fontdb::Database::new();
     font_db.load_system_fonts();
@@ -478,7 +117,7 @@ fn create_font_face() -> ResultType<Face<'static>> {
     Ok(face)
 }
 
-fn create_event_loop() -> ResultType<()> {
+pub(super) fn create_event_loop() -> ResultType<()> {
     let face = match create_font_face() {
         Ok(face) => Some(face),
         Err(err) => {
@@ -492,28 +131,11 @@ fn create_event_loop() -> ResultType<()> {
         .with_title("RustDesk whiteboard")
         .with_transparent(true)
         .with_always_on_top(true)
+        .with_skip_taskbar(true)
         .with_decorations(false);
 
-    use tao::dpi::{PhysicalPosition, PhysicalSize};
     let mut final_size = None;
-    if let Ok(displays) = crate::server::display_service::try_get_displays() {
-        let mut min_x = i32::MAX;
-        let mut min_y = i32::MAX;
-        let mut max_x = i32::MIN;
-        let mut max_y = i32::MIN;
-
-        for display in displays {
-            let (x, y) = (display.origin().0 as i32, display.origin().1 as i32);
-            let (w, h) = (display.width() as i32, display.height() as i32);
-            min_x = min_x.min(x);
-            min_y = min_y.min(y);
-            max_x = max_x.max(x + w);
-            max_y = max_y.max(y + h);
-        }
-
-        let (x, y) = (min_x, min_y);
-        let (w, h) = ((max_x - min_x) as u32, (max_y - min_y) as u32);
-
+    if let Ok((x, y, w, h)) = super::server::get_displays_rect() {
         if w > 0 && h > 0 {
             final_size = Some(PhysicalSize::new(w, h));
             window_builder = window_builder
@@ -526,11 +148,6 @@ fn create_event_loop() -> ResultType<()> {
     } else {
         window_builder =
             window_builder.with_fullscreen(Some(tao::window::Fullscreen::Borderless(None)));
-    }
-
-    #[cfg(any(target_os = "windows", target_os = "linux"))]
-    {
-        window_builder = window_builder.with_skip_taskbar(true);
     }
 
     let window = Arc::new(window_builder.build::<(String, CustomEvent)>(&event_loop)?);
